feat: added node and node badge to log info modal
